More R CMD check error fixes

+++ src/putty/sshbn.h
@@ -22,7 +22,9 @@
 
 #if defined _RTOOLS & defined _WIN64
 typedef unsigned __int64 __uint64_t;
-typedef unsigned __int128 __uint128_t;
+// WCC:del
+// typedef unsigned __int128 __uint128_t;
+typedef unsigned long long __uint128_t;
 #endif
 
 #if defined __SIZEOF_INT128__
